multisell int -> long
git-svn-id: http://svn.l2jserver.com/trunk@3153 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/entity/Castle.java
@@ -1470,7 +1470,7 @@ public class Castle
 			con = L2DatabaseFactory.getInstance().getConnection();
 			
 			statement = con.prepareStatement(CASTLE_UPDATE_CROP);
-			statement.setInt(1, (int)amount);
+			statement.setLong(1, amount);
 			statement.setInt(2, cropId);
 			statement.setInt(3, getCastleId());
 			statement.setInt(4, period);
